# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Protocol Agnostic Zero Trust Identity Broker",
        "title": "Shyntr Identity Hub API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/admin/management/tenants/{id}/scopes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all authorization scopes associated with a specific tenant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scopes"
                ],
                "summary": "List Scopes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of scopes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Scope"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve scopes",
                        "schema": {
                            "$ref": "#/definitions/payload.AppError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new authorization scope for a specific tenant.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scopes"
                ],
                "summary": "Create Scope",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Scope configuration payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Scope"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Scope created successfully",
                        "schema": {
                            "$ref": "#/definitions/model.Scope"
                        }
                    },
                    "400": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/payload.AppError"
                        }
                    },
                    "500": {
                        "description": "Failed to create scope",
                        "schema": {
                            "$ref": "#/definitions/payload.AppError"
                        }
                    }
                }
            }
        },
        "/admin/management/tenants/{id}/scopes/{scope_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the details of a specific authorization scope within a tenant boundary.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scopes"
                ],
                "summary": "Get Scope",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Scope ID",
                        "name": "scope_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Scope details",
                        "schema": {
                            "$ref": "#/definitions/model.Scope"
                        }
                    },
                    "404": {
                        "description": "Scope not found",
                        "schema": {
                            "$ref": "#/definitions/payload.AppError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing authorization scope configuration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scopes"
                ],
                "summary": "Update Scope",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Scope ID",
                        "name": "scope_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Scope update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Scope"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status: updated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/payload.AppError"
                        }
                    },
                    "500": {
                        "description": "Failed to update scope",
                        "schema": {
                            "$ref": "#/definitions/payload.AppError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes an authorization scope from a specific tenant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scopes"
                ],
                "summary": "Delete Scope",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Scope ID",
                        "name": "scope_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Failed to delete scope",
                        "schema": {
                            "$ref": "#/definitions/payload.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Scope": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "claims": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_system": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "payload.AppError": {
            "type": "object",
            "properties": {
                "status_code": {
                    "type": "integer",
                    "example": 400
                },
                "user_message": {
                    "type": "string",
                    "example": "Invalid input provided"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}